requirements.txt
django==3.2.11
arrow==1.2.1
black==20.8b1
boto3==1.20.37
# CacheControl==0.12.10
celery==5.2.3
codacy-coverage==1.3.11
# contextlib2==21.6.0
coverage==6.2
# cssselect==1.1.0
# distlib==0.3.3
# distro==1.6.0
django-cors-headers==3.11.0
django-phonenumber-field==6.0.0
django-ses==2.3.1
django-settings-export==1.2.1
django-storages==1.12.3
# docker==5.0.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-compound-fields==2.0.0
drf-yasg==1.20.0
html5lib==1.1
# importlib-metadata==4.10.1
# ipaddr==2.2.0
# lockfile==0.12.2
# lxml==4.7.1
openpyxl==3.0.9
pdfkit==1.0.0
phonenumbers==8.12.41
# Pillow==9.0.0
pip-check==2.7
pipdeptree==2.2.1
# progress==1.6
psycopg2-binary==2.9.3
pytest-django==4.5.2
python-dotenv==0.19.2
# python-memcached==1.59
pytoml==0.1.21
redis==4.1.1
# retrying==1.3.3
sentry-sdk==1.5.2
# setuptools==60.5.0
# uWSGI==2.0.20
# wheel==0.37.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
mysql
mysqlclient